refactor Input,Output 의존성 관계 제거
